merge all recent PRs into master (#174)
* issue/1823: utilise Adapt.wait

* Resolves #2001, in preparation for v3 Spoor now listens for 'tracking… (#172)

* Resolves #2001, in preparation for v3 Spoor now listens for 'tracking:complete'

This update to Spoor is intended for release with v3 of the framework.  The following properties have been removed:
- _requireCourseCompleted
- _requireAssessmentPassed

These are both replaced with corresponding properties on the framework's config.json, which is used to trigger the 'tracking:complete' event.

* Corrected issue with onTrackingComplete()

The call to Adapt.offlineStorage.set() should have been using the completionStatus variable.

* Updates following code review

- Removed submitAssessmentFailed() function -- this is now handled by onTrackingComplete()
- Removed references to main.html from README.md

* Issue #1970, switches 'configModel:dataLoaded' for 'offlineStorage:pr… (#173)

* Issue #1970, switches 'configModel:dataLoaded' for 'offlineStorage:prepare'

This allows using other plugins to implement offlineStorage providers, e.g. xAPI.

* Defaulted <title> tag in HTML to "Adapt"

This is consistent with the framework issue #1886.

* Tidy up following code review

- Moved 'app:dataReady' to a singe line anonymous function
- Removed check for removeEventListener() -- this was only required for IE8 support