remove section about editing manifest file
now the course title is being added dynamically it's not really needed. most LMSes don't seem to care whether the identifiers are unique to the SCO or not - and AT users can't make those sorts of edits anyway - so we may as well drop that part too as it's irrelevant to all but a few people.